Add dry_run mode for myrm package and make some tweaks

+++ myrm/rmlib.py
@@ -16,9 +16,10 @@ __all__ = (
 )
 
 
-def mkdir(path: str) -> None:
+def mkdir(path: str, dryrun: bool = False) -> None:
     try:
-        os.makedirs(path)
+        if not dryrun:
+            os.makedirs(path)
         logger.info("Directory '%s' was created.", path)
     except OSError as err:
         if not (err.errno == errno.EEXIST and os.path.isdir(path)):
@@ -28,9 +29,10 @@ def mkdir(path: str) -> None:
             sys.exit(getattr(err, "errno", errno.EPERM))
 
 
-def rm(path: str) -> None:
+def rm(path: str, dryrun: bool = False) -> None:
     try:
-        os.remove(path)
+        if not dryrun:
+            os.remove(path)
         logger.info("Item '%s' was deleted from the current machine.", path)
     except OSError as err:
         logger.error("Item's path can not be deleted from the current machine.")
@@ -39,7 +41,7 @@ def rm(path: str) -> None:
         sys.exit(getattr(err, "errno", errno.EPERM))
 
 
-def rmdir(path: str) -> None:
+def rmdir(path: str, dryrun: bool = False) -> None:
     try:
         content = os.walk(path, topdown=False)
     except OSError as err:
@@ -51,13 +53,14 @@ def rmdir(path: str) -> None:
     for top, dirs, nondirs in content:
         # Step -- 1.
         for name in nondirs:
-            rm(os.path.join(top, name))
+            rm(os.path.join(top, name), dryrun)
 
         try:
             # Step -- 2.
             for name in dirs:
                 abspath = os.path.join(top, name)
-                os.rmdir(abspath)
+                if not dryrun:
+                    os.rmdir(abspath)
                 logger.info("Folder '%s' was deleted from the current machine.", abspath)
         except OSError as err:
             logger.error("The determined path can't be deleted from the current machine.")
@@ -66,7 +69,8 @@ def rmdir(path: str) -> None:
             sys.exit(getattr(err, "errno", errno.EPERM))
 
     try:
-        os.rmdir(path)
+        if not dryrun:
+            os.rmdir(path)
     except OSError as err:
         logger.error("The determined path can't be deleted from the current machine.")
         logger.debug("An unexpected error occurred at this program runtime:", exc_info=True)
@@ -76,9 +80,10 @@ def rmdir(path: str) -> None:
         logger.info("Directory '%s' was deleted from the current machine.", path)
 
 
-def mv(src: str, dst: str) -> None:
+def mv(src: str, dst: str, dryrun: bool = False) -> None:
     try:
-        os.rename(src, dst)
+        if not dryrun:
+            os.rename(src, dst)
         logger.info("Item '%s' was moved to '%s' as a destinational path.", src, dst)
     except OSError as err:
         logger.error("The determined item can't be moved to the destinational path.")
@@ -87,7 +92,7 @@ def mv(src: str, dst: str) -> None:
         sys.exit(getattr(err, "errno", errno.EPERM))
 
 
-def mvdir(src: str, dst: str) -> None:
+def mvdir(src: str, dst: str, dryrun: bool = False) -> None:
     try:
         content = os.walk(src, topdown=False)
     except OSError as err:
@@ -96,15 +101,15 @@ def mvdir(src: str, dst: str) -> None:
         # Stop this program runtime and return the exit status code.
         sys.exit(getattr(err, "errno", errno.EPERM))
 
-    for top, _, nondirs in content:
-        abspath = top.replace(src, dst)
+    if not dryrun:
+        for top, _, nondirs in content:
+            abspath = top.replace(src, dst)
 
-        # Step -- 1.
-        mkdir(abspath)
+            # Step -- 1.
+            mkdir(abspath, dryrun)
 
-        # Step -- 2.
-        for name in nondirs:
-            mv(os.path.join(top, name), os.path.join(abspath, name))
-
-    rmdir(src)
+            # Step -- 2.
+            for name in nondirs:
+                mv(os.path.join(top, name), os.path.join(abspath, name), dryrun)
+        rmdir(src, dryrun)
     logger.info("Directory '%s' was moved to '%s' as a destinational path.", src, dst)
